+++ ali/common-lootjs/src/main/java/com/yanny/ali/lootjs/Utils.java
@@ -1,9 +1,19 @@
 package com.yanny.ali.lootjs;
 
 import com.almostreliable.lootjs.core.LootType;
+import com.almostreliable.lootjs.core.entry.ItemLootEntry;
 import com.mojang.logging.LogUtils;
+import com.yanny.aci.api.RangeValue;
+import com.yanny.ali.api.IDataNode;
+import com.yanny.ali.api.IServerUtils;
+import com.yanny.ali.lootjs.node.ItemStackNode;
+import com.yanny.ali.plugin.common.NodeUtils;
 import net.minecraft.resources.ResourceLocation;
+import net.minecraft.world.level.storage.loot.entries.LootItem;
+import net.minecraft.world.level.storage.loot.functions.LootItemFunction;
+import net.minecraft.world.level.storage.loot.predicates.LootItemCondition;
 import org.jetbrains.annotations.NotNull;
+import org.jetbrains.annotations.Nullable;
 import org.slf4j.Logger;
 
 import java.lang.reflect.Field;
@@ -54,5 +64,14 @@ public class Utils {
             case SHEARING -> (r) -> r.getPath().startsWith("shearing/");
             case GENERIC -> (r) -> true;
         };
+    }
+
+    @NotNull
+    public static IDataNode getEntry(IServerUtils utils, ItemLootEntry entry, float chance, List<LootItemFunction> functions, List<LootItemCondition> conditions, @Nullable RangeValue preservedCount) {
+        LootItem vanillaEntry = entry.getVanillaEntry();
+        List<LootItemCondition> allConditions = NodeUtils.getAllConditions(vanillaEntry, conditions);
+        List<LootItemFunction> allFunctions = NodeUtils.getAllFunctions(vanillaEntry, functions);
+
+        return new ItemStackNode(utils, vanillaEntry.item.value().getDefaultInstance(), chance, allFunctions, allConditions, preservedCount);
     }
 }