+++ ali/common-lootjs/src/main/java/com/yanny/ali/lootjs/server/LootJsGenericTooltipUtils.java
@@ -6,6 +6,7 @@ import com.almostreliable.lootjs.core.filters.ItemFilterImpl;
 import com.yanny.aci.tooltip.TooltipBuilder;
 import com.yanny.ali.api.IServerUtils;
 import com.yanny.ali.language.Lang;
+import com.yanny.ali.plugin.server.MissingTooltipUtils;
 import net.minecraft.advancements.critereon.MinMaxBounds;
 import net.minecraft.core.component.DataComponentType;
 import net.minecraft.tags.TagKey;
@@ -37,7 +38,7 @@ public class LootJsGenericTooltipUtils {
         } else if (predicate == ItemFilter.DAMAGEABLE) {
             return TooltipBuilder.value("DAMAGEABLE");
         } else if (predicate == ItemFilter.DAMAGED) {
-            return TooltipBuilder.value("ENCHANTABLE");
+            return TooltipBuilder.value("DAMAGED");
         } else if (predicate == ItemFilter.ENCHANTED) {
             return TooltipBuilder.value("ENCHANTED");
         } else if (predicate == ItemFilter.BLOCK_ITEM) {
@@ -90,10 +91,10 @@ public class LootJsGenericTooltipUtils {
     public static TooltipBuilder getIdFilterTooltip(IServerUtils utils, IdFilter filter) {
         return switch (filter) {
             case IdFilter.ByLocation byLocation -> utils.getValueTooltip(utils, byLocation.location());
-            case IdFilter.ByPattern byPattern -> utils.getValueTooltip(utils, byPattern.toString()).key(Lang.Value.PATTERN);
+            case IdFilter.ByPattern byPattern -> utils.getValueTooltip(utils, byPattern.pattern().pattern()).key(Lang.Value.PATTERN);
             case IdFilter.ByMod byMod -> utils.getValueTooltip(utils, byMod.mod()).key(Lang.Value.MOD);
             case IdFilter.Or or -> utils.getValueTooltip(utils, or.filters()).key(Lang.Branch.OR);
-            default -> throw new IllegalStateException("Unexpected IdFilter type: " + filter);
+            default -> MissingTooltipUtils.getMissingValueTooltip(utils, filter);
         };
     }